+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/geospatial/GeoShapeSearchSpatialOperationsIT.java
@@ -21,6 +21,8 @@ import com.stratio.cassandra.lucene.testsAT.util.CassandraUtils;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -30,6 +32,7 @@ import static com.stratio.cassandra.lucene.builder.Builder.*;
 /**
  * @author Eduardo Alonso {@literal <eduardoalonso@stratio.com>}
  */
+@RunWith(JUnit4.class)
 public class GeoShapeSearchSpatialOperationsIT extends BaseIT {
 
     private static String shape_1 = "POLYGON((0 0,1 0, 1 1,0 1,0 0))";
@@ -122,7 +125,7 @@ public class GeoShapeSearchSpatialOperationsIT extends BaseIT {
 
     @AfterClass
     public static void after() {
-        utils.dropIndex().dropTable().dropKeyspace();
+        CassandraUtils.dropKeyspaceIfNotNull(utils);
     }
 
     @Test