+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/issues/Issue69IT.java
@@ -20,12 +20,15 @@ import com.stratio.cassandra.lucene.testsAT.util.CassandraUtils;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import static org.junit.Assert.assertEquals;
 
 /**
  * @author Andres de la Pena <adelapena@stratio.com>
  */
+@RunWith(JUnit4.class)
 public class Issue69IT extends BaseIT {
 
     private static CassandraUtils utils;
@@ -50,7 +53,7 @@ public class Issue69IT extends BaseIT {
 
     @AfterClass
     public static void after() {
-        utils.dropKeyspace();
+        CassandraUtils.dropKeyspaceIfNotNull(utils);
     }
 
     @Test