+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/varia/StatelessSearchWithSkinnyRowsIT.java
@@ -62,7 +62,7 @@ public class StatelessSearchWithSkinnyRowsIT extends BaseIT {
 
     @AfterClass
     public static void after() {
-        utils.dropIndex().dropTable().dropKeyspace();
+        CassandraUtils.dropKeyspaceIfNotNull(utils);
     }
 
     @Test