+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 
+## 3.8.5 (upcoming)
+
+* Fix collection columns indexing being part of clustering key (#286)
+* Avoid unsupported usage of multicolumn mappers with collections
+* Add support for indexing map values
+* Add support for indexing map keys (#253)
+
 ## 3.8.4 (February 06, 2017)
 
 * Fix mutual exclusion on partition key during read before write (#261)
 * Ignore tombstones during read before write (#255)
 * Fix deletion of ranges of rows in a partition
 * Add virtual node based partitioner
-
 
 ## 3.8.3 (December 15, 2016)